Split invoke tasks into categories.
