+++ perma_web/tasks/__init__.py
@@ -15,7 +15,17 @@ except Exception as e:
 from .dev import run, pip_compile
 from . import dev
 
+# Special tasks
+from . import once
+from . import wacz_conversion
+from . import check_storage
+from . import merge_accounts
+
 ns = Collection()
 ns.add_task(run)
 ns.add_task(pip_compile)
 ns.add_collection(dev)
+ns.add_collection(once)
+ns.add_collection(wacz_conversion)
+ns.add_collection(check_storage)
+ns.add_collection(merge_accounts)
